requirements.txt
numpy==1.23.3
opencv-python==4.6.0.66
pdf2image==1.16.0
Pillow==9.2.0
